#fixes cleanup of queues functionality moved to utils
