+++ src/app.ts
@@ -8,8 +8,7 @@ import { customAlphabet } from "nanoid"
 import { LegacyAPI, OpenAPISchema, HTTPS_CERT, _bootstrap_db } from "./utils"
 import { ActivityScheduler } from "./utils/ActivitySchedulerJob"
 import API from "./service"
-import {SchedulerQueue} from "./utils/queue/SchedulerQueue"
-import {SchedulerReferenceQueue} from "./utils/queue/SchedulerReferenceQueue"
+import {cleanAllQueues} from "./utils/ActivitySchedulerJob"
 
 // The database connection and ID generators for repository classes.
 export const Database = nano(process.env.CDB ?? "")
@@ -25,29 +24,6 @@ app.use(cors())
 app.use(morgan(":method :url :status - :response-time ms"))
 app.use(express.urlencoded({ extended: true }))
 
-
-//clean all jobs
-async function cleanAllQueues(): Promise<any> {
-  await SchedulerQueue.clean(0, 'delayed');
-  await SchedulerQueue.clean(0, 'wait');
-  await   SchedulerQueue.clean(0, 'active');
-  await  SchedulerQueue.clean(0, 'completed');
-  await  SchedulerQueue.clean(0, 'failed');
-  let multi_1 = SchedulerQueue.multi();
-  await   multi_1.del(SchedulerQueue.toKey('repeat'));
-  await  multi_1.exec();
-
-  await SchedulerReferenceQueue.clean(0, 'delayed');
-  await SchedulerReferenceQueue.clean(0, 'wait');
-  await   SchedulerReferenceQueue.clean(0, 'active');
-  await  SchedulerReferenceQueue.clean(0, 'completed');
-  await  SchedulerReferenceQueue.clean(0, 'failed');
-  let multi_2 = SchedulerReferenceQueue.multi();    
-  await  multi_2.exec();
-
-  await SchedulerQueue.empty()
-  await SchedulerReferenceQueue.empty()
-}
 // Initialize and configure the application.
 async function main(): Promise<void> {
   console.group("Initializing LAMP API server...")
